Hacky workaround for half-finalized builds.
Metalava increments the SDK level by one when it's not "REL", so we
temporarily force the build to be "REL" while we're still in the
process of finalizing it.

This CL must be reverted as part of actually declaring "REL".

Bug: 171506470
Test: Build
(cherry picked from commit 3c5d6eafad29bd0d141e457c10e219a20e72994b)

Change-Id: I18405cdd27769598f5e7a6a968f0dd473f936f6c